Move authority parsing into rpc package
